client/requirements.txt
boto3==1.24.24
botocore==1.27.24
gTTS==2.2.4
moviepy==1.0.3
playwright==1.23.0
praw==7.6.1
prawcore~=2.3.0
pytube==12.1.0
requests>=2.28.1
rich==12.5.1
toml==0.10.2
translators==5.7.6
pyttsx3==2.90
Pillow~=9.1.1
tomlkit==0.11.4
Flask==2.2.2
spacy==3.4.1
